// EAM_CLI/src/app/models/tran/item-order.model.ts
export class ItemOrder {
  id: string | null = 'A';
  aufnr: string | null = '';
  category: string | null = 'M';
  matnr: string | null = null;
  marktxt: string | null = null;
  werks: string | null = null;
  budat: Date | null = null;
  isActive: boolean | null = true;
  menge: number | null = 0;
  meins: string | null = null;
  category2: string | null = null;
  lgort: string | null = null;
  charg: string | null = null;
  price: number | null = 0;
  dmbtr: number | null = 0;
  waers: string | null = null;
  uname: string | null = null;
  udat: Date | null = null;
}